Abstracted type recovery to remove language specific components
