+++ joern-cli/frontends/x2cpg/src/main/scala/io/joern/x2cpg/passes/frontend/impl/PythonTypeHintCallLinker.scala
@@ -1,0 +1,12 @@
+package io.joern.x2cpg.passes.frontend.impl
+
+import io.joern.x2cpg.passes.frontend.XTypeHintCallLinker
+import io.shiftleft.codepropertygraph.Cpg
+import io.shiftleft.codepropertygraph.generated.nodes.Call
+import overflowdb.traversal.Traversal
+
+class PythonTypeHintCallLinker(cpg: Cpg) extends XTypeHintCallLinker(cpg) {
+
+  override def calls: Traversal[Call] = super.calls.filterNot(c => c.name.matches("^(import).*"))
+
+}
